+++ src/client/java/com/cubiccadence/client/provider/netease/NeteaseApiClient.java
@@ -1,0 +1,5 @@
+package com.cubiccadence.client.provider.netease;
+
+public class NeteaseApiClient {
+    // TODO: wrap NetEase Cloud Music open platform HTTP calls
+}